requirements.txt
Pillow==4.0.0
xlrd==1.0.0
